agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Anaskhan", email = "anaskhanx101@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]

[project.scripts]
agent = "agent:main"
cli = "agent.cli:main"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"
